refactor: MSW files structure
