+++ src/mocks/handlers/index.js
@@ -1,0 +1,7 @@
+import { userHandlers } from "./user"
+import { toDosHandlers } from "./toDo"
+
+export const handlers = [
+  ...userHandlers,
+  ...toDosHandlers
+]
