+++ src/mocks/server.js
@@ -1,4 +1,4 @@
 import { setupServer } from 'msw/node'
-import { handlers } from './handlers'
+import { handlers } from './handlers/index'
 
 export const server = setupServer(...handlers)
